refactor: put docs on client builder methods instead of fluent builder rename: generateShapeMemberList -> generateShapeMemberDocs refactor: move generateShapeMemberList to FluentClientDecorator.kt update: appease spellchecker
